realtime_processor/requirements.txt
pika>=1.2.0
paho-mqtt>=1.6.1
pydantic>=2.11.5